feat: sonar task-6 2
