+++ .github/workflows/task6-pipeline.yml
@@ -1,0 +1,118 @@
+name: Task 6 CI/CD Pipeline
+
+on:
+  push:
+    branches: [ main, task-6 ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build-test-sonar-docker:
+    runs-on: ubuntu-latest
+    env:
+      SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+      SONAR_HOST_URL: ${{ secrets.SONAR_HOST_URL }}
+      DOCKER_REGISTRY: ${{ secrets.DOCKER_REGISTRY }}
+      DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
+      DOCKER_PASSWORD: ${{ secrets.DOCKER_PASSWORD }}
+      IMAGE_NAME: flask_app
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.10'
+
+      - name: Install dependencies
+        run: |
+          cd app
+          pip install -r requirements.txt
+          pip install pytest
+
+      - name: Run unit tests
+        run: |
+          cd app
+          pytest
+
+      - name: SonarQube Scan
+        uses: SonarSource/sonarcloud-github-action@v2
+        with:
+          projectBaseDir: app
+        env:
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+          SONAR_HOST_URL: ${{ secrets.SONAR_HOST_URL }}
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Login to Docker Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ secrets.DOCKER_REGISTRY }}
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Build and push Docker image
+        run: |
+          docker build -t $DOCKER_REGISTRY/$IMAGE_NAME:latest .
+          docker push $DOCKER_REGISTRY/$IMAGE_NAME:latest
+
+  deploy-verify:
+    needs: build-test-sonar-docker
+    runs-on: ubuntu-latest
+    env:
+      KUBECONFIG_DATA: ${{ secrets.KUBECONFIG_DATA }}
+      DOCKER_REGISTRY: ${{ secrets.DOCKER_REGISTRY }}
+      IMAGE_NAME: flask_app
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up kubectl
+        uses: azure/setup-kubectl@v4
+        with:
+          version: 'latest'
+
+      - name: Set up Helm
+        uses: azure/setup-helm@v4
+        with:
+          version: 'latest'
+
+      - name: Configure kubeconfig
+        run: |
+          echo "$KUBECONFIG_DATA" | base64 -d > kubeconfig
+          export KUBECONFIG=$PWD/kubeconfig
+
+      - name: Helm deploy
+        run: |
+          helm upgrade --install flask-app helm/flask-app --set image.repository=$DOCKER_REGISTRY/$IMAGE_NAME --set image.tag=latest
+
+      - name: Pobierz publiczny IP EC2 (node)
+        id: get_ip
+        run: |
+          export KUBECONFIG=$PWD/kubeconfig
+          NODE_IP=$(kubectl get nodes -o jsonpath='{.items[0].status.addresses[?(@.type=="ExternalIP")].address}')
+          echo "NODE_IP=$NODE_IP" >> $GITHUB_ENV
+
+      - name: Pobierz NodePort aplikacji
+        id: get_port
+        run: |
+          export KUBECONFIG=$PWD/kubeconfig
+          APP_PORT=$(kubectl -n default get svc flask-app -o jsonpath='{.spec.ports[0].nodePort}')
+          echo "APP_PORT=$APP_PORT" >> $GITHUB_ENV
+
+      - name: App verification (curl)
+        run: |
+          sleep 10
+          echo "Testuję aplikację na http://$NODE_IP:$APP_PORT"
+          curl -f http://$NODE_IP:$APP_PORT || exit 1
+
+      - name: Notify on success
+        if: success()
+        run: echo "Deployment succeeded!"
+
+      - name: Notify on failure
+        if: failure()
+        run: echo "Deployment failed!"
